fixed two issues in SoaringController and renamed it into AP_Soaring
