+++ libraries/AP_Soaring/AP_Soaring.h
